+++ src/gnomonVisualization/gnomonVisualizations/gnomonImage/gnomonVisualizationImageChannelBlending.h
@@ -55,8 +55,8 @@ public slots:
     void onXY(void) override;
     void onXZ(void) override;
     void onYZ(void) override;
-    void onSliceChanged(double) override;
-    void onSliceOrientationChanged(double) override;
+    void onSliceChanged(int) override;
+    void onSliceOrientationChanged(int) override;
     void onTimeChanged(double) override;
 
 public:
